Duplicating Cabins

+++ 17-the-wild-oasis/src/services/apiCabins.js
@@ -51,12 +51,13 @@ export async function createEditCabin(newCabin, id) {
   }
 
   // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
   // 3. Delete the cabin IF there was an error in uploading image
-
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.log(storageError);
